nfc-test/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
adafruit-circuitpython-pn532==2.3.18
adafruit-blinka==8.22.2
RPi.GPIO==0.7.1